fix!: always return BigInt for INT64 columns (COUNT NULL fix) (#5)
* fix: add fallback type detection for COUNT() and aggregates

- Add error logging when row.column() fails for known TypeCodes
- Add fallback type detection when TypeCode is unrecognized (0)
- Try Int64, Float64, String, Bool in sequence before returning null
- Fixes COUNT() returning NULL instead of actual value

* style: cargo fmt

* chore: add cargo fmt comment to AGENTS.md

* fix!: always return BigInt for INT64 columns

BREAKING CHANGE: INT64 values now always return as i64 (BigInt), never i32 (Int).

- Spanner only has INT64, so returning BigInt ensures type compatibility
- Fix COUNT() returning NULL when used with FromQueryResult (i64 field)
- Fix subquery FROM parsing for SeaORM's count() method
- Add find_top_level_from() to handle nested subqueries

Entity fields should use i64 instead of i32 for integer columns.

* test: update entities and tests for BigInt-only INT64

- Change i32 fields to i64 in test entities (user, product, all_types)
- Update integer literals in type_tests.rs
- Add COUNT + GROUP BY aggregate tests
- Replace .count() with .all().len() due to SeaORM u64 limitation

* docs: update Integer Types section for i64-only recommendation

+++ tests/entity/all_types.rs
@@ -11,8 +11,8 @@ pub struct Model {
 
     pub int64_val: i64,
     pub int64_nullable: Option<i64>,
-    pub int32_val: i32,
-    pub int32_nullable: Option<i32>,
+    pub int32_val: i64,
+    pub int32_nullable: Option<i64>,
 
     pub float64_val: f64,
     pub float64_nullable: Option<f64>,
